Moved CustomResendVerificationSerializer, CustomVerifyEmailSerializer, CustomSetPasswordSerializer to utils and renamed them to MyResendVerificationSerializer, MyVerifyEmailSerializer, MySetPasswordSerializer; Updated docstrings
